+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,8 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Rank;
-import blackjack.domain.card.Shape;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
